move EnvSelectMenu
